+++ demisto_sdk/commands/common/constants.py
@@ -825,7 +825,8 @@ DIR_TO_PREFIX = {
 
 ENTITY_NAME_SEPARATORS = [' ', '_', '-']
 
-DELETED_YML_FIELDS_BY_DEMISTO = ['fromversion', 'toversion', 'alt_dockerimages', 'script.dockerimage45', 'tests']
+DELETED_YML_FIELDS_BY_DEMISTO = ['fromversion', 'toversion', 'alt_dockerimages', 'script.dockerimage45', 'tests',
+                                 'defaultclassifier', 'defaultmapperin', 'defaultmapperout']
 
 DELETED_JSON_FIELDS_BY_DEMISTO = ['fromVersion', 'toVersion']