+++ demisto_sdk/commands/format/update_classifier.py
@@ -11,8 +11,14 @@ from demisto_sdk.commands.format.update_generic_json import BaseUpdateJSON
 
 
 class BaseClassifierJSONFormat(BaseUpdateJSON, ABC):
-    def __init__(self, input: str = '', output: str = '', path: str = '', from_version: str = '',
-                 no_validate: bool = False, verbose: bool = False):
+    def __init__(self,
+                 input: str = '',
+                 output: str = '',
+                 path: str = '',
+                 from_version: str = '',
+                 no_validate: bool = False,
+                 verbose: bool = False,
+                 **kwargs):
         super().__init__(input=input, output=output, path=path, from_version=from_version, no_validate=no_validate,
                          verbose=verbose)